refactor: 불필요한 객체 제거
